backend/requirements.txt
# --- Web framework ---
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# --- Data / Validation ---
pydantic==2.9.2
pydantic-settings==2.6.1

# --- Database ---
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
greenlet==3.1.1

# --- Cache / Queue ---
redis==5.1.1
celery==5.4.0
kombu==5.4.2

# --- Auth / Crypto ---
PyJWT==2.9.0
webauthn==2.3.0
cryptography==43.0.3
passlib==1.7.4

# --- HTTP clients ---
httpx==0.27.2
aiohttp==3.10.10
requests==2.32.3

# --- Networking / Recon ---
dnspython==2.7.0

# --- ML / IDS ---
joblib==1.4.2
numpy==2.1.2
scikit-learn==1.5.2

# --- Observability ---
structlog==24.4.0
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0

# --- Rules / STIX helpers ---
PyYAML==6.0.2

# --- Utilities ---
python-dotenv==1.0.1
orjson==3.10.11
tenacity==9.0.0
